Changes on the User model. Moved list of orders in the user class. Added Add/Remove order in the user base class. other minor changes.
