+++ DeliverIT/Common/Validator.cs
@@ -14,6 +14,7 @@ namespace DeliverIT.Common
                 throw new ArgumentNullException();
             }
         }
+
         public static void ValidateUserInfo(string input, int min, int max, string message)
         {
             if (string.IsNullOrEmpty(input) || input.Length < min || input.Length > max)
