+++ DeliverIT/Models/Contracts/User/ICourier.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using DeliverIT.Contracts;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -9,7 +10,6 @@ namespace DeliverIT.Models.Contracts.User
     public interface ICourier : IUser
     {
 
-        IList<Order> Deliveries { get; set; }
         int Id { get; }
         double AllowedVolume { get; set; }
         double AllowedWeight { get; set; }
