+++ DeliverIT/Models/Users/Couriers/Abstract/Courier.cs
@@ -2,20 +2,17 @@
 using DeliverIT.Common;
 using DeliverIT.Models.Users.Abstract;
 using DeliverIT.Models.Contracts.User;
+using DeliverIT.Contracts;
 
 namespace DeliverIT.Models.Users.Couriers.Abstract
 {
     public abstract class Courier : User, ICourier
     {
-        private IList<Order> deliveries;
         private static int id = 0;
         private double allowedVolume;
         private double allowedWeight;
 
-        public IList<Order> Deliveries { get => deliveries; set => deliveries = value; }
-        public double AllowedVolume { get => allowedVolume; set => allowedVolume = value; }
-        public double AllowedWeight { get => allowedWeight; set => allowedWeight = value; }
-        public int Id { get => id; }
+
 
         public Courier(string firstName, string lastName, string email, string phoneNumber,
                         int years, Address address, GenderType gender,
@@ -25,26 +22,35 @@ namespace DeliverIT.Models.Users.Couriers.Abstract
             id++;
             this.AllowedVolume = allowedVolume;
             this.AllowedWeight = allowedWeight;
-            this.Deliveries = new List<Order>();
         }
+
+        public double AllowedVolume { get => allowedVolume; set => allowedVolume = value; }
+        public double AllowedWeight { get => allowedWeight; set => allowedWeight = value; }
+        public int Id { get => id; }
 
         //method used for allowed volume and weight orders
         public virtual bool CanCarry()
         {
             bool canCarry = false;
 
-            foreach (var delivery in Deliveries)
+            foreach (var order in this.OrdersList)
             {
-                if (delivery.Product.Weight <= this.AllowedWeight &&
-                    delivery.Product.Volume <= this.AllowedVolume)
+                
+                if (order.Product.Weight <= this.AllowedWeight &&
+                    order.Product.Volume <= this.AllowedVolume)
                 {
-                    this.AllowedWeight -= delivery.Product.Weight;
-                    this.AllowedVolume -= delivery.Product.Volume;
+                    this.AllowedWeight -= order.Product.Weight;
+                    this.AllowedVolume -= order.Product.Volume;
                     canCarry = true;
                 }
             }
 
             return canCarry;
+        }
+
+        public void DisplayAllOrders()
+        {
+
         }
 
         public override string ToString()
